+++ node/src/service.rs
@@ -14,25 +14,34 @@
 // You should have received a copy of the GNU General Public License
 // along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
+use async_io::Timer;
 use cumulus_network::build_block_announce_validator;
 use cumulus_service::{
 	prepare_node_config, start_collator, start_full_node, StartCollatorParams, StartFullNodeParams,
 };
 use fc_consensus::FrontierBlockImport;
+use sc_consensus_manual_seal::{run_manual_seal, EngineCommand, ManualSealParams};
 use fc_rpc_core::types::{FilterPool, PendingTransactions};
+use futures::{Stream, StreamExt};
 use moonbeam_runtime::{opaque::Block, RuntimeApi};
 use polkadot_primitives::v0::CollatorPair;
 use sc_client_api::BlockchainEvents;
 use sc_executor::native_executor_instance;
 pub use sc_executor::NativeExecutor;
-use sc_service::{Configuration, PartialComponents, Role, TFullBackend, TFullClient, TaskManager};
-use sp_core::{Pair, H160};
+use sc_service::{error::Error as ServiceError,Configuration, PartialComponents, Role, TFullBackend, TFullClient, TaskManager};
+use sp_core::{Pair, H160, H256};
 use sp_runtime::traits::BlakeTwo256;
 use sp_trie::PrefixedMemoryDB;
 use std::{
+	time::Duration,
 	collections::{BTreeMap, HashMap},
 	sync::{Arc, Mutex},
 };
+use crate::{
+	cli::Sealing,
+	inherents::build_inherent_data_providers,
+};
+
 // Our native executor instance.
 native_executor_instance!(
 	pub Executor,
@@ -51,6 +60,7 @@ type FullBackend = TFullBackend<Block>;
 pub fn new_partial(
 	config: &Configuration,
 	author: Option<H160>,
+	mock_inherents: bool,
 ) -> Result<
 	PartialComponents<
 		FullClient,
@@ -64,9 +74,9 @@ pub fn new_partial(
 			Option<FilterPool>,
 		),
 	>,
-	sc_service::Error,
+	ServiceError,
 > {
-	let inherent_data_providers = crate::inherents::build_inherent_data_providers(author, false)?;
+	let inherent_data_providers = build_inherent_data_providers(author, mock_inherents)?;
 
 	let (client, backend, keystore_container, task_manager) =
 		sc_service::new_full_parts::<Block, RuntimeApi, Executor>(&config)?;
@@ -87,6 +97,10 @@ pub fn new_partial(
 
 	let frontier_block_import = FrontierBlockImport::new(client.clone(), client.clone(), true);
 
+	// We use the cumulus import queue here regardless of whether we're running a real parachain or
+	// the dev service. Typically manual seal would use the manual seal import queue. In reality,
+	// both are pretty minimal. The cumulus one checks the inherents, so we use it.
+	// https://github.com/paritytech/substrate/issues/8164
 	let import_queue = cumulus_consensus::import_queue::import_queue(
 		client.clone(),
 		frontier_block_import.clone(),
@@ -95,7 +109,7 @@ pub fn new_partial(
 		registry,
 	)?;
 
-	let params = PartialComponents {
+	Ok(PartialComponents {
 		backend,
 		client,
 		import_queue,
@@ -105,9 +119,7 @@ pub fn new_partial(
 		inherent_data_providers,
 		select_chain: (),
 		other: (frontier_block_import, pending_transactions, filter_pool),
-	};
-
-	Ok(params)
+	})
 }
 
 /// Start a node with the given parachain `Configuration` and relay chain `Configuration`.
@@ -143,7 +155,7 @@ where
 			},
 		)?;
 
-	let params = new_partial(&parachain_config, author_id)?;
+	let params = new_partial(&parachain_config, author_id, false)?;
 
 	let client = params.client.clone();
 	let backend = params.backend.clone();
@@ -214,7 +226,7 @@ where
 
 	// Spawn Frontier EthFilterApi maintenance task.
 	if filter_pool.is_some() {
-		use futures::StreamExt;
+		// use futures::StreamExt;
 		// Each filter is allowed to stay in the pool for 100 blocks.
 		const FILTER_RETAIN_THRESHOLD: u64 = 100;
 		task_manager.spawn_essential_handle().spawn(
@@ -239,7 +251,7 @@ where
 	// Spawn Frontier pending transactions maintenance task (as essential, otherwise we leak).
 	if pending_transactions.is_some() {
 		use fp_consensus::{ConsensusLog, FRONTIER_ENGINE_ID};
-		use futures::StreamExt;
+		// use futures::StreamExt;
 		use sp_runtime::generic::OpaqueDigestItemId;
 
 		const TRANSACTION_RETAIN_THRESHOLD: u64 = 5;
@@ -353,4 +365,225 @@ pub async fn start_node(
 		|_| Default::default(),
 	)
 	.await
+}
+
+/// Builds a new development service. This service uses manual seal, and mocks
+/// the parachain inherent.
+pub fn new_dev(
+	config: Configuration,
+	sealing: Sealing,
+	author_id: Option<H160>,
+) -> Result<TaskManager, ServiceError> {
+	let sc_service::PartialComponents {
+		client,
+		backend,
+		mut task_manager,
+		import_queue,
+		keystore_container,
+		select_chain: _,
+		transaction_pool,
+		inherent_data_providers,
+		other: (block_import, pending_transactions, filter_pool),
+	} = new_partial(&config, author_id, true)?;
+
+	let (network, network_status_sinks, system_rpc_tx, network_starter) =
+		sc_service::build_network(sc_service::BuildNetworkParams {
+			config: &config,
+			client: client.clone(),
+			transaction_pool: transaction_pool.clone(),
+			spawn_handle: task_manager.spawn_handle(),
+			import_queue,
+			on_demand: None,
+			block_announce_validator_builder: None,
+		})?;
+
+	if config.offchain_worker.enabled {
+		sc_service::build_offchain_workers(
+			&config,
+			backend.clone(),
+			task_manager.spawn_handle(),
+			client.clone(),
+			network.clone(),
+		);
+	}
+
+	let role = config.role.clone();
+	let prometheus_registry = config.prometheus_registry().cloned();
+	let is_authority = role.is_authority();
+	let subscription_task_executor =
+		sc_rpc::SubscriptionTaskExecutor::new(task_manager.spawn_handle());
+	let mut command_sink = None;
+
+	if role.is_authority() {
+		let env = sc_basic_authorship::ProposerFactory::new(
+			task_manager.spawn_handle(),
+			client.clone(),
+			transaction_pool.clone(),
+			prometheus_registry.as_ref(),
+		);
+
+		let commands_stream: Box<dyn Stream<Item = EngineCommand<H256>> + Send + Sync + Unpin> =
+			match sealing {
+				Sealing::Instant => {
+					Box::new(
+						// This bit cribbed from the implementation of instant seal.
+						transaction_pool
+							.pool()
+							.validated_pool()
+							.import_notification_stream()
+							.map(|_| EngineCommand::SealNewBlock {
+								create_empty: false,
+								finalize: false,
+								parent_hash: None,
+								sender: None,
+							}),
+					)
+				}
+				Sealing::Manual => {
+					let (sink, stream) = futures::channel::mpsc::channel(1000);
+					// Keep a reference to the other end of the channel. It goes to the RPC.
+					command_sink = Some(sink);
+					Box::new(stream)
+				}
+				Sealing::Interval(millis) => Box::new(StreamExt::map(
+					Timer::interval(Duration::from_millis(millis)),
+					|_| EngineCommand::SealNewBlock {
+						create_empty: true,
+						finalize: false,
+						parent_hash: None,
+						sender: None,
+					},
+				)),
+			};
+
+		// Typically the longest chain rule is constructed in `new_partial`, but the full service,
+		// does not use one. So instead we construct our longest chain rule here.
+		let select_chain = sc_consensus::LongestChain::new(backend.clone());
+
+		task_manager.spawn_essential_handle().spawn_blocking(
+			"authorship_task",
+			run_manual_seal(ManualSealParams {
+				block_import,
+				env,
+				client: client.clone(),
+				pool: transaction_pool.pool().clone(),
+				commands_stream,
+				select_chain,
+				consensus_data_provider: None,
+				inherent_data_providers,
+			}),
+		);
+	}
+
+	let rpc_extensions_builder = {
+		let client = client.clone();
+		let pool = transaction_pool.clone();
+		let network = network.clone();
+		let pending = pending_transactions.clone();
+		let filter_pool = filter_pool.clone();
+		Box::new(move |deny_unsafe, _| {
+			let deps = crate::rpc::FullDeps {
+				client: client.clone(),
+				pool: pool.clone(),
+				graph: pool.pool().clone(),
+				deny_unsafe,
+				is_authority,
+				network: network.clone(),
+				pending_transactions: pending.clone(),
+				filter_pool: filter_pool.clone(),
+				command_sink: command_sink.clone(),
+			};
+			crate::rpc::create_full(deps, subscription_task_executor.clone())
+		})
+	};
+
+	sc_service::spawn_tasks(sc_service::SpawnTasksParams {
+		network,
+		client: client.clone(),
+		keystore: keystore_container.sync_keystore(),
+		task_manager: &mut task_manager,
+		transaction_pool: transaction_pool.clone(),
+		rpc_extensions_builder,
+		on_demand: None,
+		remote_blockchain: None,
+		backend,
+		network_status_sinks,
+		system_rpc_tx,
+		config,
+	})?;
+
+	// Spawn Frontier EthFilterApi maintenance task.
+	if filter_pool.is_some() {
+		// Each filter is allowed to stay in the pool for 100 blocks.
+		const FILTER_RETAIN_THRESHOLD: u64 = 100;
+		task_manager.spawn_essential_handle().spawn(
+			"frontier-filter-pool",
+			client
+				.import_notification_stream()
+				.for_each(move |notification| {
+					if let Ok(locked) = &mut filter_pool.clone().unwrap().lock() {
+						let imported_number: u64 = notification.header.number as u64;
+						for (k, v) in locked.clone().iter() {
+							let lifespan_limit = v.at_block + FILTER_RETAIN_THRESHOLD;
+							if lifespan_limit <= imported_number {
+								locked.remove(&k);
+							}
+						}
+					}
+					futures::future::ready(())
+				}),
+		);
+	}
+
+	// Spawn Frontier pending transactions maintenance task (as essential, otherwise we leak).
+	if pending_transactions.is_some() {
+		use fp_consensus::{ConsensusLog, FRONTIER_ENGINE_ID};
+		use sp_runtime::generic::OpaqueDigestItemId;
+
+		const TRANSACTION_RETAIN_THRESHOLD: u64 = 5;
+		task_manager.spawn_essential_handle().spawn(
+			"frontier-pending-transactions",
+			client
+				.import_notification_stream()
+				.for_each(move |notification| {
+					if let Ok(locked) = &mut pending_transactions.clone().unwrap().lock() {
+						// As pending transactions have a finite lifespan anyway
+						// we can ignore MultiplePostRuntimeLogs error checks.
+						let mut frontier_log: Option<_> = None;
+						for log in notification.header.digest.logs.iter().rev() {
+							let log = log.try_to::<ConsensusLog>(OpaqueDigestItemId::Consensus(
+								&FRONTIER_ENGINE_ID,
+							));
+							if log.is_some() {
+								frontier_log = log;
+								break;
+							}
+						}
+
+						let imported_number: u64 = notification.header.number as u64;
+
+						if let Some(ConsensusLog::EndBlock {
+							block_hash: _,
+							transaction_hashes,
+						}) = frontier_log
+						{
+							// Retain all pending transactions that were not
+							// processed in the current block.
+							locked.retain(|&k, _| !transaction_hashes.contains(&k));
+						}
+						locked.retain(|_, v| {
+							// Drop all the transactions that exceeded the given lifespan.
+							let lifespan_limit = v.at_block + TRANSACTION_RETAIN_THRESHOLD;
+							lifespan_limit > imported_number
+						});
+					}
+					futures::future::ready(())
+				}),
+		);
+	}
+
+	log::info!("Development Service Ready");
+
+	network_starter.start_network();
+	Ok(task_manager)
 }
